Fix: Moving the DBWarehouse in the SQLiteRepository class.
